Update after review
Change param 'admin' to 'user' in StaffForm.
Delete blocks in confirm_promote.html.
Refactor tests:
- Create test_group with tests from test_permissions which does not tests access rights (tests for views, forms).
- Create test_staff with tests from test_permissions which does not tests access rights (tests for views, forms).
