+++ saleor/dashboard/staff/views.py
@@ -29,7 +29,7 @@ def staff_details(request, pk):
     queryset = User.objects.filter(is_staff=True)
     staff_member = get_object_or_404(queryset, pk=pk)
     form = StaffForm(request.POST or None, instance=staff_member,
-                     admin=request.user)
+                     user=request.user)
     if form.is_valid():
         form.save()
         msg = pgettext_lazy(
